Use shared layout for profile pages Also move user profiles pages to same location
