+++ src/assets/screens/layout/Layout.jsx
@@ -3,8 +3,8 @@ import { useState } from 'react'
 import { BrowserRouter, Navigate, Route, Routes } from 'react-router-dom'
 import PropTypes from 'prop-types';
 import ProtectedRoute from '../../../routes/ProtectedRoute'
-import ExhibitionCardCreate from '../../components/ExhibitionCard/ExhibitionCardCreate'
-import PostDetail from '../../components/Post/PostDetail'
+import ExhibitionCardCreate from '../userProfile/ExhibitionCardCreate'
+import PostDetail from '../Post/PostDetail'
 import Login from '../../components/VerificationPages/LoginPage'
 import SignUp from '../../components/VerificationPages/SignUpPage'
 import AdminDashboard from '../Admin/AdminDashboard'
@@ -21,7 +21,7 @@ import NewsPage from '../newsPage/NewsPage'
 import UserProfile from '../userProfile/userProfile'
 import UserProfileAddPost from '../userProfile/userProfileAddPost'
 import UserProfilePosts from '../userProfile/userProfilePosts'
-import ProductCardCreate from '/src/assets/components/ProductCard/ProductCardCreate'
+import ProductCardCreate from '../userProfile/ProductCardCreate'
 import styles from '/src/styles/layout/Layout.module.scss'
 import ArtTermsPage from '../ArtTerms/ArtTermsPage'
 import ArtTermsFilteredPage from '../ArtTerms/ArtTermsFilteredPage'
