+++ src/routes/Routes.jsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom'
-import ExhibitionCardCreate from '../assets/components/ExhibitionCard/ExhibitionCardCreate.jsx'
-import PostDetail from '../assets/components/Post/PostDetail.jsx'
-import ProductCartCreate from '../assets/components/ProductCard/ProductCardCreate.jsx'
+import ExhibitionCardCreate from '../assets/screens/userProfile/ExhibitionCardCreate.jsx'
+import PostDetail from '../assets/screens/Post/PostDetail.jsx'
+import ProductCartCreate from '../assets/screens/userProfile/ProductCardCreate.jsx'
 import Login from '../assets/components/VerificationPages/LoginPage.jsx'
 import SignUp from '../assets/components/VerificationPages/SignUpPage.jsx'
 import AdminDashboard from '../assets/screens/Admin/AdminDashboard.jsx'
